Fixed warning

+++ include/str/parse.h
@@ -21,7 +21,7 @@ DECLARE_ENUM_WITH_TYPE(u8, clt, ParseErrorCode, SUCCESS, INVALID_FMT, OUT_OF_RAN
 namespace clt::meta
 {
   template<Integral T>
-  inline constexpr u64 max_digits10_v = clt::ceil(clt::log10(std::numeric_limits<T>::max())) + std::is_signed_v<T>;
+  inline constexpr u64 max_digits10_v = static_cast<u64>(clt::ceil(clt::log10(std::numeric_limits<T>::max()))) + std::is_signed_v<T>;
 }
 
 namespace clt::str
